% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas-package.R
\docType{package}
\name{christmas-package}
\alias{christmas}
\alias{christmas-package}
\title{christmas: Generation of Different Animated Christmas Cards}
\description{
Generation of different Christmas cards, most of them being animated. Most of the cards can be generated in three languages (English, Catalan and Spanish). The collection started in 2009.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sites.google.com/view/josebarrera/}
}

}
\author{
\strong{Maintainer}: Jose Barrera-Gomez \email{jose.barrera@uab.cat} (\href{https://orcid.org/0000-0002-2688-6036}{ORCID})

}
\keyword{internal}
