% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas.R
\docType{package}
\name{christmas}
\alias{christmas}
\alias{christmas-package}
\title{christmas: A package for creating a number of Christmas cards.}
\description{
This is a collection of Christmas cards, mostly of them being animated. The name of each card includes the year in which it was created.
}
\author{
Jose Barrera-Gomez \email{jose.barrera@isglobal.org}
}
