% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowman_card.R
\name{snowman_card}
\alias{snowman_card}
\title{Snowman card}
\usage{
snowman_card(
  filepath,
  message,
  greeting = "Merry Christmas",
  n_snowflakes = 100,
  seed = 2512
)
}
\arguments{
\item{filepath}{File path to save image to}

\item{message}{Individual message}

\item{greeting}{Greeting message. Default \code{"Merry Christmas"}}

\item{n_snowflakes}{Number of snowflakes. Default 100.}

\item{seed}{Random seed. Defaul 2512.}
}
\value{
PNG file
}
\description{
Create a snowman Christmas card
}
