% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2008tree.R
\name{xmas2008tree}
\alias{xmas2008tree}
\title{Christmas card 2008.}
\usage{
xmas2008tree(year = 2009, language = c("english", "spanish",
  "catalan"), seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2009}.}

\item{language}{Language to be used in the card. One of \code{c("english", "spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including a random tree.
}
\description{
Christmas card 2008: a random Christmas tree.
}
\examples{
\donttest{
xmas2008tree()
xmas2008tree(year = 2020, language = "catalan")
xmas2008tree(year = 2020, language = "spanish", seed = 666)
}
}
\author{
Jose Barrera-Gomez.
}
