% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2009caganer.R
\name{xmas2009caganer}
\alias{xmas2009caganer}
\title{Christmas card 2009.}
\usage{
xmas2009caganer(seed = NULL)
}
\arguments{
\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot with a caganer.
}
\description{
Christmas card 2009: a 'caganer'. Caganer is is a figurine depicted in the act of defecation appearing in nativity scenes in Catalonia and neighbouring areas with Catalan culture such as Andorra, Valencia, and Northern Catalonia (in southern France). It is most popular and widespread in these areas, but can also be found in other areas of Spain (Murcia), Portugal, and southern Italy (Naples). Further details can be found at \url{https://en.wikipedia.org/wiki/Caganer}. This caganer won the annual contest of sustainable caganers at CREAL (now ISGlobal, \url{https://www.isglobal.org/en/}) in 2009.
}
\examples{
\donttest{
xmas2009caganer()
}
}
\author{
Jose Barrera-Gomez.
}
