% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2010text.R
\name{xmas2010text}
\alias{xmas2010text}
\title{Christmas card 2010.}
\usage{
xmas2010text(year = 2011, language = c("english", "spanish",
  "catalan"), seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2011}.}

\item{language}{Language to be used in the card. One of \code{c("english", "spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including the repetition of the same message randomly allocated.
}
\description{
Christmas card 2010: random allocation of repetitions of the Christmas message.
}
\examples{
\donttest{
xmas2010text()
xmas2010text(year = 2020, language = "spanish", seed = 666)
}
}
\author{
Jose Barrera-Gomez.
}
