% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2011health.R
\name{xmas2011health}
\alias{xmas2011health}
\title{Christmas card 2011.}
\usage{
xmas2011health(seed = NULL)
}
\arguments{
\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including boxplots and OR estimates.
}
\description{
Christmas card 2011: some epidemiological jokes about potential effects of Christmas on health.
}
\examples{
\donttest{
xmas2011health()
}
}
\author{
Jose Barrera-Gomez.
}
