% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2012snowkoch.R
\name{xmas2012snowkoch}
\alias{xmas2012snowkoch}
\title{Christmas card 2012.}
\usage{
xmas2012snowkoch(year = 2013, seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2013}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including Koch snowflakes.
}
\description{
Christmas card 2012: Koch snowflakes (https://en.wikipedia.org/wiki/Koch_snowflake) are used to print the Christmas message in three languages (English, Catalan and Spanish).
}
\examples{
\donttest{
xmas2012snowkoch()
}
}
\author{
Jose Barrera-Gomez.
}
