% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2013neon.R
\name{xmas2013neon}
\alias{xmas2013neon}
\title{Christmas card 2013.}
\usage{
xmas2013neon(language = c("english", "spanish", "catalan"),
  nflash = 30, pause = 0.1, seed = NULL)
}
\arguments{
\item{language}{Language to be used in the card. One of \code{c("english", "spanish", "catalan")}. Default is \code{"english"}.}

\item{nflash}{Number of flashes shown in the neon sign. Default is 30.}

\item{pause}{Pause, in seconds, between two consecutive flashes shown in the neon sign. Default is 0.1.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including a message and then a simulations of a neon sign.
}
\description{
Christmas card 2013: a neon sign.
}
\examples{
\donttest{
xmas2013neon()
xmas2013neon(language = "catalan", nflash = 10, pause = 0.3)
xmas2013neon(language = "spanish", nflash = 8, pause = 1)
xmas2013neon(nflash = 100, pause = 0.02)
}
}
\author{
Jose Barrera-Gomez.
}
