% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2014snowman.R
\name{xmas2014snowman}
\alias{xmas2014snowman}
\title{Christmas card 2014.}
\usage{
xmas2014snowman(year = 2015, language = c("english", "spanish",
  "catalan"), seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2015}.}

\item{language}{Language to be used in the card. One of \code{c("english", "spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including a snowman.
}
\description{
Christmas card 2014: a snowman.
}
\examples{
\donttest{
xmas2014snowman()
xmas2014snowman(year = 2020, language = "catalan")
xmas2014snowman(year = 2020, language = "spanish")
}
}
\author{
Jose Barrera-Gomez.
}
