% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2015flag.R
\name{xmas2015flag}
\alias{xmas2015flag}
\title{Christmas card 2015.}
\usage{
xmas2015flag(year = 2016, language = c("english", "spanish",
  "catalan"), seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2016}.}

\item{language}{Language to be used in the card. One of \code{c("english", "spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including a message in a flag.
}
\description{
Christmas card 2015: a flag
}
\examples{
\donttest{
xmas2015flag()
xmas2015flag(year = 2020, language = "catalan")
xmas2015flag(year = 2020, language = "spanish")
}
}
\author{
Jose Barrera-Gomez.
}
