% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas2016epi.R
\name{xmas2016epi}
\alias{xmas2016epi}
\title{Christmas card 2016.}
\usage{
xmas2016epi(year = 2017, seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2017}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including a tree made of nice words.
}
\description{
Christmas card 2016: a statistical/epidemiological Christmas tree. This tree is built using a random subset of statistical and epidemiological words.
}
\examples{
\donttest{
xmas2016epi()
xmas2016epi(year = 2020)
}
}
\author{
Jose Barrera-Gomez.
}
