% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmascaganer.R
\name{xmascaganer}
\alias{xmascaganer}
\title{Caganer.}
\usage{
xmascaganer(
  year = 2010,
  language = c("english", "spanish", "catalan"),
  seed = NULL
)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2010}.}

\item{language}{Language to be used in the card. One of \code{c("english",
"spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot with a caganer.
}
\description{
A Catalan 'caganer' (2009 card). Caganer is a figurine depicted
  in the act of defecation appearing in nativity scenes in Catalonia and
  neighbouring areas with Catalan culture such as Andorra, Valencia, and
  Northern Catalonia (in southern France). It is most popular and widespread
  in these areas, but can also be found in other areas of Spain (Murcia),
  Portugal, and southern Italy (Naples). Further details can be found at
  \url{https://en.wikipedia.org/wiki/Caganer}. This caganer won the annual
  contest of sustainable caganers at CREAL (now ISGlobal,
  \url{https://www.isglobal.org/en/}) in 2009.
}
\examples{
\donttest{
xmascaganer()
}
}
\author{
Jose Barrera-Gomez.
}
