% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmasepitree.R
\name{xmasepitree}
\alias{xmasepitree}
\title{Christmas epitree.}
\usage{
xmasepitree(year = 2017, seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2017}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot including a tree made of nice words.
}
\description{
A statistical/epidemiological Christmas tree (2016 card). This
  tree is built using a random subset of statistical and epidemiological
  words.
}
\examples{
\donttest{
xmasepitree()
}
}
\author{
Jose Barrera-Gomez.
}
