% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmasflag.R
\name{xmasflag}
\alias{xmasflag}
\title{Christmas flag.}
\usage{
xmasflag(
  year = 2016,
  language = c("english", "spanish", "catalan"),
  seed = NULL
)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2016}.}

\item{language}{Language to be used in the card. One of \code{c("english",
"spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot including a message in a flag.
}
\description{
Christmas flag (2015 card). A flag with a Christmas message.
}
\examples{
\donttest{
xmasflag()
xmasflag(year = 2020, language = "catalan")
xmasflag(year = 2020, language = "spanish")
}
}
\author{
Jose Barrera-Gomez.
}
