% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmasfractaltree.R
\name{xmasfractaltree}
\alias{xmasfractaltree}
\title{Fractal Christmas tree.}
\usage{
xmasfractaltree(year = 2025, language = c("english", "spanish", "catalan"))
}
\arguments{
\item{year}{Year to be printed. Default is \code{2025}.}

\item{language}{Language to be used in the card. One of \code{c("english", "spanish", "catalan")}. Default is \code{"english"}.}
}
\value{
A Christmas card plot including a fractal Christmas tree.
}
\description{
The tree fractal structure was adapted from
  (https://blogs.sas.com/content/iml/2012/12/14/a-fractal-christmas-tree.html).
  A preliminary translation to R was written by Alejandra Cabana (UAB).
}
\examples{
\donttest{
xmasfractaltree()
}
}
\author{
Jose Barrera-Gomez.
}
