% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmashealth.R
\name{xmashealth}
\alias{xmashealth}
\title{Christmas risks.}
\usage{
xmashealth(year = 2012, seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2012}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot including boxplots and OR estimates.
}
\description{
Christmas risks (2011 card). Some epidemiological jokes about
  potential effects of Christmas on health.
}
\examples{
\donttest{
xmashealth()
}
}
\author{
Jose Barrera-Gomez.
}
