% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmashelix.R
\name{xmashelix}
\alias{xmashelix}
\title{Christmas helix.}
\usage{
xmashelix(
  year = 2024,
  language = c("english", "spanish", "catalan"),
  totalsecs = NULL,
  seed = NULL
)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2024}.}

\item{language}{Language to be used in the card. One of \code{c("english",
"spanish", "catalan")}. Default is \code{"english"}.}

\item{totalsecs}{Approximate total time, in seconds, to create the card.
Default is \code{NULL} (and it is internally set).}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot including a message in a helix.
}
\description{
Christmas helix (2023 card). A helix with a Christmas message.
}
\examples{
\donttest{
xmashelix(year = 2024, language = "catalan", totalsecs = 1)
}
}
\author{
Jose Barrera-Gomez.
}
