% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmaspicasso.R
\name{xmaspicasso}
\alias{xmaspicasso}
\title{Picasso's pigeon.}
\usage{
xmaspicasso(
  year = 2023,
  language = c("english", "spanish", "catalan"),
  seed = NULL
)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2023}.}

\item{language}{Language to be used in the card. One of \code{c("english", "spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including an approximation to Picasso's pigeon.
}
\description{
Approximation to Picasso's pigeon (2022 card).
}
\examples{
\donttest{
xmaspicasso()
xmaspicasso(year = 2023, language = "catalan")
}
}
\author{
Jose Barrera-Gomez.
}
