% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmassnowkoch.R
\name{xmassnowkoch}
\alias{xmassnowkoch}
\title{Koch snowflakes.}
\usage{
xmassnowkoch(year = 2013, seed = NULL)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2013}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot including Koch snowflakes.
}
\description{
Koch snowflakes (2012 card). Koch snowflakes
  (https://en.wikipedia.org/wiki/Koch_snowflake) are used to print the
  Christmas message in three languages (English, Catalan and Spanish).
}
\examples{
\donttest{
xmassnowkoch()
}
}
\author{
Jose Barrera-Gomez.
}
