% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmassnowman.R
\name{xmassnowman}
\alias{xmassnowman}
\title{Christmas snowman.}
\usage{
xmassnowman(
  year = 2015,
  language = c("english", "spanish", "catalan"),
  seed = NULL
)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2015}.}

\item{language}{Language to be used in the card. One of \code{c("english", "spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no seed).}
}
\value{
A Christmas card plot including a snowman.
}
\description{
Christmas snowman (2014 card).
}
\examples{
\donttest{
xmassnowman()
xmassnowman(year = 2020, language = "catalan")
xmassnowman(year = 2020, language = "spanish")
}
}
\author{
Jose Barrera-Gomez.
}
