% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmastext.R
\name{xmastext}
\alias{xmastext}
\title{Christmas message.}
\usage{
xmastext(
  year = 2011,
  language = c("english", "spanish", "catalan"),
  seed = NULL
)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2011}.}

\item{language}{Language to be used in the card. One of \code{c("english",
"spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot including the repetition of the same message
  randomly allocated.
}
\description{
A random repetition of a message (2010 card). Random allocation
  of repetitions of the Christmas message.
}
\examples{
\donttest{
xmastext(year = 2011, language = "spanish", seed = 666)
}
}
\author{
Jose Barrera-Gomez.
}
