% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmastree.R
\name{xmastree}
\alias{xmastree}
\title{Naive Christmas tree.}
\usage{
xmastree(
  year = 2009,
  language = c("english", "spanish", "catalan"),
  seed = NULL
)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2009}.}

\item{language}{Language to be used in the card. One of \code{c("english",
"spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot including a random tree.
}
\description{
A random naive Christmas tree (2008 card).
}
\examples{
\donttest{
xmastree(year = 2020, language = "catalan", seed = 666)
}
}
\author{
Jose Barrera-Gomez.
}
