% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmastreewire.R
\name{xmastreewire}
\alias{xmastreewire}
\title{Wire Christmas tree.}
\usage{
xmastreewire(
  year = 2022,
  language = c("english", "spanish", "catalan"),
  seed = NULL
)
}
\arguments{
\item{year}{Year to be printed. Default is \code{2022}.}

\item{language}{Language to be used in the card. One of \code{c("english",
"spanish", "catalan")}. Default is \code{"english"}.}

\item{seed}{Seed for reproducibility of the card. Default is \code{NULL} (no
seed).}
}
\value{
A Christmas card plot including a random wire tree.
}
\description{
A random wire Christmas tree (2021 card).
}
\examples{
\donttest{
xmastreewire()
xmastreewire(year = 2020, language = "catalan")
xmastreewire(year = 2020, language = "spanish", seed = 666)
}
}
\author{
Jose Barrera-Gomez.
}
