% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_santa.R
\name{secret_santa}
\alias{secret_santa}
\title{Secret Santa}
\usage{
secret_santa(names, seed = 2024)
}
\arguments{
\item{names}{Vector of names.}

\item{seed}{Random seed.}
}
\value{
Data frame of names
}
\description{
Create Secret Santa allocations
}
